[!!!][TASK] Remove `additionalJavaScriptSubmit` from FormEngine
With RequireJS in place, it is possible to minimize the amount of inline
JavaScript in the backend. This patch removes FormEngine's
`additionalJavaScriptSubmit` feature that registers additional submit
handlers and allows to abort the form submission.

This feature is rarely used at all, but can be achieved using a RequireJS
module that registers a callback in `TYPO3/CMS/Backend/DocumentSaveActions`.

Since the feature is gone, a lot of global code within TBE_EDITOR is
removed as well.

Resolves: #88667
Releases: master
Change-Id: I92f26aa52cafab5df30a706ea95f1e1702a4a778
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60861
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>