[FEATURE] Add email address to installation process
New installations now ask for the email address as well, so
this can be used later on to e.g. notify the admin if somebody
logged-in (Warning email address).

The same feature is available in the maintenance module of
the Install Tool in the "Create Administrative User" card.

Resolves: #89227
Releases: master
Change-Id: Iafca7043543d5e299fb121d96d9b3b9caa375bc9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60889
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>